[RFC] store: add compression support (#2)
Implement compression on store side. If compression is enabled data is compressed before being written to segment files on disk and decompressed on read. The code which deals with records remains the same, it doesn't affected by this patch.

Now `gzip` and `zstandard` are implemented, but other implementations can also be easily added.

To enable compression `-compression` flag for `oklog store` or `oklog ingeststore` should be provided.

oklog/oklog#134